[connectors] move activemq connectors to new model #1667 (modules rename)
