+++ app/connector/activemq/pom.xml
@@ -16,12 +16,12 @@
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-    <parent>
-        <groupId>io.syndesis.connector</groupId>
-        <artifactId>connector-parent</artifactId>
-        <version>1.3-SNAPSHOT</version>
-        <relativePath>../../../pom.xml</relativePath>
-    </parent>
+  <parent>
+    <groupId>io.syndesis.connector</groupId>
+    <artifactId>connector-parent</artifactId>
+    <version>1.3-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
+  </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>connector-activemq</artifactId>
@@ -65,13 +65,13 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>connector-verifier</artifactId>
+      <artifactId>connector-support-verifier</artifactId>
       <optional>true</optional>
       <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>io.syndesis.connector</groupId>
-      <artifactId>connector-utils</artifactId>
+      <artifactId>connector-support-util</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -103,7 +103,7 @@
 
     <dependency>
       <groupId>io.syndesis.connector</groupId>
-      <artifactId>connector-test</artifactId>
+      <artifactId>connector-support-test</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
@@ -144,7 +144,7 @@
       </plugin>
       <plugin>
         <groupId>${project.groupId}</groupId>
-        <artifactId>connector-maven-plugin</artifactId>
+        <artifactId>connector-support-maven-plugin</artifactId>
         <version>${project.version}</version>
         <executions>
           <execution>
