+++ app/connector/activemq/src/main/java/io/syndesis/connector/activemq/ActiveMQConnector.java
