+++ app/connector/activemq/src/main/resources/META-INF/spring.factories
