+++ app/connector/activemq/src/main/resources/META-INF/syndesis/connector/activemq.json
