+++ app/connector/activemq/src/test/java/io/syndesis/connector/activemq/ActiveMQConnectorTestSupport.java
@@ -17,11 +17,11 @@ package io.syndesis.connector.activemq;
 
 import java.util.function.Consumer;
 
-import io.syndesis.connector.test.ConnectorTestSupport;
-import io.syndesis.model.action.ConnectorAction;
-import io.syndesis.model.connection.Connector;
-import io.syndesis.model.integration.Step;
-import io.syndesis.model.integration.StepKind;
+import io.syndesis.connector.support.test.ConnectorTestSupport;
+import io.syndesis.common.model.action.ConnectorAction;
+import io.syndesis.common.model.connection.Connector;
+import io.syndesis.common.model.integration.Step;
+import io.syndesis.common.model.integration.StepKind;
 import org.apache.activemq.junit.EmbeddedActiveMQBroker;
 import org.junit.Rule;
 import org.junit.rules.TestName;
@@ -50,7 +50,7 @@ public abstract class ActiveMQConnectorTestSupport extends ConnectorTestSupport 
         final Step.Builder builder = new Step.Builder()
             .stepKind(StepKind.endpoint)
             .action(action)
-            .connection(new io.syndesis.model.connection.Connection.Builder()
+            .connection(new io.syndesis.common.model.connection.Connection.Builder()
                 .connector(connector)
                 .putConfiguredProperty("brokerUrl", broker.getVmURL())
                 .build());
